requirements.txt
langchain_community==0.2.1
langchain_core==0.2.1
langchain_google_genai==1.0.5
langchain_text_splitters==0.2.0
protobuf==5.27.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
SQLAlchemy_Utils==0.41.2
streamlit==1.33.0
streamlit_option_menu==0.3.12